sllin: Use separate spinlock for LIN frame cache and processing protection.
The sl->linfr_lock spinlock is used with spin_lock_irqsave
exclusivelly and it is safe to use it in all kinds
of contexts (thread, bottom-halves or even IRQ handlers).
The sl->lock protects network device related data
and outside of sll_xmit or other bottom-halves callbacks
has to be locked with spin_lock_bh variant.

Signed-off-by: Pavel Pisa <pisa@cmp.felk.cvut.cz>