sllin: Avoid race in netdev initialization
There is a netdev_dbg before the call to wait_event_killable in
in sllin_kwthread. As the worker thread is started before the
netdev is initialized, this is a race.

This shouldn't be the end of the world seeing as netdev_dbg handles
uninitialized netdevs. Still, initializing the device before starting
the thread avoids this altogether.